requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-discover-runner==1.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22
django-froala-editor==3.2.2
django-ipware==3.0.2
django-nine==0.2.4
django-redis==4.12.1
django-rest-elasticsearch==0.4.2
django-role-permissions==3.1.1
django-webpack-loader==0.7.0
djangorestframework==3.12.4
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
Faker==8.1.0
idna==2.10
ipaddress==1.0.23
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.4
gunicorn==20.1.0